Consolidated redundant code into a single preview tab.
