+++ src/megameklab/com/ui/Mek/MainUI.java
@@ -40,8 +40,8 @@ import megamek.common.TripodMech;
 import megameklab.com.ui.MegaMekLabMainUI;
 import megameklab.com.ui.Mek.tabs.BuildTab;
 import megameklab.com.ui.Mek.tabs.EquipmentTab;
-import megameklab.com.ui.Mek.tabs.PreviewTab;
 import megameklab.com.ui.Mek.tabs.StructureTab;
+import megameklab.com.ui.tabs.PreviewTab;
 import megameklab.com.util.MenuBarCreator;
 import megameklab.com.util.UnitUtil;
